Decouple FeedItemUtil and PlaybackStatus
