+++ app/src/main/java/de/danoeh/antennapod/adapter/actionbutton/ItemActionButton.java
@@ -8,10 +8,10 @@ import androidx.annotation.StringRes;
 import android.view.View;
 
 import de.danoeh.antennapod.core.service.download.DownloadService;
+import de.danoeh.antennapod.core.util.PlaybackStatus;
 import de.danoeh.antennapod.model.feed.FeedItem;
 import de.danoeh.antennapod.model.feed.FeedMedia;
 import de.danoeh.antennapod.core.preferences.UserPreferences;
-import de.danoeh.antennapod.core.util.FeedItemUtil;
 
 public abstract class ItemActionButton {
     FeedItem item;
@@ -40,7 +40,7 @@ public abstract class ItemActionButton {
         }
 
         final boolean isDownloadingMedia = DownloadService.isDownloadingFile(media.getDownload_url());
-        if (FeedItemUtil.isCurrentlyPlaying(media)) {
+        if (PlaybackStatus.isCurrentlyPlaying(media)) {
             return new PauseActionButton(item);
         } else if (item.getFeed().isLocalFeed()) {
             return new PlayLocalActionButton(item);
