+++ app/src/main/java/de/danoeh/antennapod/menuhandler/FeedItemMenuHandler.java
@@ -23,6 +23,7 @@ import de.danoeh.antennapod.core.sync.SynchronizationSettings;
 import de.danoeh.antennapod.core.sync.queue.SynchronizationQueueSink;
 import de.danoeh.antennapod.core.util.FeedItemUtil;
 import de.danoeh.antennapod.core.util.IntentUtils;
+import de.danoeh.antennapod.core.util.PlaybackStatus;
 import de.danoeh.antennapod.core.util.ShareUtils;
 import de.danoeh.antennapod.dialog.ShareDialog;
 import de.danoeh.antennapod.model.feed.FeedItem;
@@ -52,7 +53,7 @@ public class FeedItemMenuHandler {
             return false;
         }
         final boolean hasMedia = selectedItem.getMedia() != null;
-        final boolean isPlaying = hasMedia && FeedItemUtil.isPlaying(selectedItem.getMedia());
+        final boolean isPlaying = hasMedia && PlaybackStatus.isPlaying(selectedItem.getMedia());
         final boolean isInQueue = selectedItem.isTagged(FeedItem.TAG_QUEUE);
         final boolean fileDownloaded = hasMedia && selectedItem.getMedia().fileExists();
         final boolean isFavorite = selectedItem.isTagged(FeedItem.TAG_FAVORITE);
