+++ core/src/main/java/de/danoeh/antennapod/core/util/FeedItemUtil.java
@@ -75,19 +75,4 @@ public class FeedItemUtil {
         int smartMarkAsPlayedSecs = UserPreferences.getSmartMarkAsPlayedSecs();
         return media.getDuration() > 0 && media.getPosition() >= media.getDuration() - smartMarkAsPlayedSecs * 1000;
     }
-
-    /**
-     * Reads playback preferences to determine whether this FeedMedia object is
-     * currently being played and the current player status is playing.
-     */
-    public static boolean isCurrentlyPlaying(FeedMedia media) {
-        return isPlaying(media) && PlaybackService.isRunning
-                && ((PlaybackPreferences.getCurrentPlayerStatus() == PlaybackPreferences.PLAYER_STATUS_PLAYING));
-    }
-
-    public static boolean isPlaying(FeedMedia media) {
-        return PlaybackPreferences.getCurrentlyPlayingMediaType() == FeedMedia.PLAYABLE_TYPE_FEEDMEDIA
-                && media != null
-                && PlaybackPreferences.getCurrentlyPlayingFeedMediaId() == media.getId();
-    }
 }
